Move some static file management to webpack
