+++ frontend/package.json
@@ -7,8 +7,7 @@
     "dep-graph": "madge -c -i \"dep-graph.png\" ./src/scripts",
     "build": "npx gulp build",
     "build:live": "npx gulp build-production",
-    "build:webpack": "webpack --config webpack.config.js",
-    "start:dev": "npm run build && concurrently --kill-others \"npx gulp watch\" \"firebase serve --only hosting\"",
+    "start:dev": "sh ./scripts/build-serve.sh",
     "deploy:live": "npm run build:live && firebase deploy -P live --only hosting",
     "tsc": "tsc"
   },
@@ -29,6 +28,7 @@
     "babel-loader": "^8.2.3",
     "buffer": "^6.0.3",
     "circular-dependency-plugin": "5.2.2",
+    "copy-webpack-plugin": "10.2.4",
     "dart-sass": "^1.25.0",
     "del": "^6.0.0",
     "eslint-webpack-plugin": "3.1.1",
@@ -37,6 +37,7 @@
     "gulp-concat": "^2.6.1",
     "gulp-eslint-new": "^1.3.0",
     "gulp-sass": "^5.0.0",
+    "html-minimizer-webpack-plugin": "3.5.0",
     "html2canvas": "^1.4.1",
     "madge": "5.0.1",
     "stream-browserify": "^3.0.0",
