+++ frontend/scripts/build-serve.sh
@@ -1,0 +1,4 @@
+npm run build &
+firebase serve --only hosting &
+npx gulp watch &
+wait
