+++ frontend/scripts/json-validation.js
